refactor: Participant 추상 클래스를 추가하여 Dealer, Player 메서드 중복 개선
